+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 #### 🐞 Fixes
 
 - Fixed Bash-based hooks being generated with the wrong path separators on Windows.
+- Fixed an issue where an inherited task with merge strategy "replace" will accidentally remove task
+  args, deps, and env in inheriting tasks.
 
 ## 1.29.0